feat: 추가된 색상 적용
